Fixed reactions implementation and refactored feed page
